pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "arggraph"
version = "0.1.0"
description = "Framework for representing, querying, and visualizing scientific argument graphs"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.0",
    "networkx>=3.0",
    "svgwrite>=1.4",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "mypy>=1.5",
]

[project.urls]
Homepage = "https://github.com/YOUR_USERNAME/arggraph"
Repository = "https://github.com/YOUR_USERNAME/arggraph"
Issues = "https://github.com/YOUR_USERNAME/arggraph/issues"

[tool.hatch.build.targets.wheel]
packages = ["argviz"]
artifacts = ["argviz/schema/*.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
